pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[project]
name = "phongthuyso_python"
version = "0.1.0"
description = "Hệ thống phân tích phong thủy số học qua Google ADK"
authors = [
    {name = "Phong Thủy Số", email = "phongthuynumerology@example.com"},
]
readme = "README.md"
requires-python = ">=3.11"
license = "MIT License"

[tool.poetry.dependencies]
python = "^3.11"
google-adk = ">=0.2.0"
google-generativeai = ">=0.3.1"
fastapi = ">=0.104.0"
uvicorn = ">=0.23.0"
pydantic = ">=2.5.0"
python-dotenv = ">=1.0.0"
gunicorn = ">=21.2.0"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pytest = ">=7.4.0"
black = ">=23.9.0"
flake8 = ">=6.1.0"
pytest-cov = ">=4.1.0" 